Extract replaceVideoTags into a shared utility
