+++ app/utils/getDocsPage.js
@@ -1,11 +1,10 @@
+import fs from 'node:fs/promises';
 import matter from 'gray-matter';
+import hljs from 'highlight.js';
 import { Marked } from 'marked';
 import { markedHighlight } from 'marked-highlight';
 import etag from './etag.server';
-
-import hljs from 'highlight.js';
-
-import fs from 'node:fs/promises';
+import { replaceVideoTags } from './replaceVideoTags';
 
 const DOCS_PATH = 'docs';
 
@@ -32,20 +31,4 @@ export default async function getDocsPage(name) {
     etag: etag(markdown),
     content: replaceVideoTags(marked.parse(markdown)),
   };
-}
-
-function replaceVideoTags(html) {
-  while (html.includes('{{<video')) {
-    const start = html.indexOf('{{<video');
-    const end = html.indexOf('}}', start);
-    const videoTag = html.substring(start, end + 2);
-
-    const src = videoTag.match(/src="([^"]+)"/)[1];
-    const _html = html.replace(
-      videoTag,
-      `<video controls><source src="${src}" type="video/mp4">Your browser does not support video</video>`,
-    );
-  }
-
-  return _html;
 }
